requirements.txt
torch
torchaudio
torchvision
transformers
accelerate
openai
python-dotenv
tiktoken
flask-cors